Work on fixing Renderpass

+++ sources/code/Plugins/GraphicsVulkan/VulkanPipeline.cpp
@@ -172,7 +172,7 @@ namespace Grindstone {
 
 			VkPipelineLayoutCreateInfo pipelineLayoutInfo = {};
 			pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-			pipelineLayoutInfo.setLayoutCount = layouts.size();
+			pipelineLayoutInfo.setLayoutCount = static_cast<uint32_t>(layouts.size());
 			pipelineLayoutInfo.pSetLayouts = layouts.data();
 			pipelineLayoutInfo.pushConstantRangeCount = 0;
